+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - Changes to crash logs generation:
   - Now all crashes are saved into the single file per game session
   - "MOP_Crash" file doesn't have game details anymore
+- Renamed "Run in background" to "Run game in background"
 - (MSCLoader) Updated to MSCLoader 1.2.6
 - (Mod Loader Pro) Renamed "CHANGE RESOLUTION" to "RESOLUTION"